+++ lib/webserver/apihandlers.py
@@ -4,10 +4,12 @@ from urllib.parse import unquote
 
 from aiohttp import web
 from aiohttp_basicauth import BasicAuthMiddleware
+from attr import field
 from lib.config import Config
 from lib.serverstate import ServerState
 from lib.socket import ClientManager
 from lib.util import safe_json_dumps
+from lib.util.util import implement_fields_query
 from lib.webserver.auth import EnodoAuth
 from lib.webserver.basehandler import BaseHandler
 
@@ -46,7 +48,8 @@ class ApiHandlers:
 
     @classmethod
     @EnodoAuth.auth.required
-    async def get_single_monitored_series(cls, request):
+    @implement_fields_query
+    async def get_single_monitored_series(cls, request, fields=None):
         """Returns all details
 
         Args:
@@ -57,13 +60,14 @@ class ApiHandlers:
         """
         series_name = unquote(request.match_info['series_name'])
         data, status = await BaseHandler.resp_get_single_monitored_series(
-            series_name)
+            series_name, fields=fields)
         return web.json_response(
             data, dumps=safe_json_dumps, status=status)
 
     @classmethod
     @EnodoAuth.auth.required
-    async def get_all_series_output(cls, request):
+    @implement_fields_query
+    async def get_all_series_output(cls, request, fields=None):
         """Returns forecast data of a specific series.
 
         Args:
@@ -73,79 +77,23 @@ class ApiHandlers:
             _type_: _description_
         """
         series_name = unquote(request.match_info['series_name'])
-        fields = None
-        if "fields" in request.rel_url.query:
-            fields = urllib.parse.unquote(
-                request.rel_url.query['fields'])
-            if fields == "":
-                fields = None
-            if fields is not None:
-                fields = fields.split(",")
         only_future = False
         if "forecastFutureOnly" in request.rel_url.query:
             val = request.rel_url.query['forecastFutureOnly']
             if val == "1" or val.lower() == "true":
                 only_future = True
+        types = None
+        if "types" in request.rel_url.query:
+            types_str = request.rel_url.query['types']
+            if types_str != "":
+                types = types_str.split(",")
+                if len(types) < 1:
+                    types = None
         return web.json_response(
             data=await BaseHandler.resp_get_all_series_output(
-                series_name, fields=fields, forecast_future_only=only_future),
+                series_name, fields=fields, forecast_future_only=only_future,
+                types=types),
             dumps=safe_json_dumps)
-
-    @classmethod
-    @EnodoAuth.auth.required
-    async def get_series_forecast(cls, request):
-        """Returns forecast data of a specific series.
-
-        Args:
-            request (Request): aiohttp request
-
-        Returns:
-            _type_: _description_
-        """
-        series_name = unquote(request.match_info['series_name'])
-        only_future = False
-        if "forecastFutureOnly" in request.rel_url.query:
-            val = request.rel_url.query['forecastFutureOnly']
-            if val == "1" or val.lowercase() == "true":
-                only_future = True
-
-        return web.json_response(
-            data=await BaseHandler.resp_get_series_forecasts(
-                series_name, only_future), dumps=safe_json_dumps)
-
-    @classmethod
-    @EnodoAuth.auth.required
-    async def get_series_anomalies(cls, request):
-        """Returns anomalies data of a specific series.
-
-        Args:
-            request (Request): aiohttp request
-
-        Returns:
-            _type_: _description_
-        """
-        series_name = unquote(request.match_info['series_name'])
-
-        return web.json_response(
-            data=await BaseHandler.resp_get_series_anomalies(
-                series_name), dumps=safe_json_dumps)
-
-    @classmethod
-    @EnodoAuth.auth.required
-    async def get_series_static_rules_hits(cls, request):
-        """Returns static rules hits of a specific series.
-
-        Args:
-            request (Request): aiohttp request
-
-        Returns:
-            _type_: _description_
-        """
-        series_name = unquote(request.match_info['series_name'])
-
-        return web.json_response(
-            data=await BaseHandler.resp_get_series_static_rules_hits(
-                series_name), dumps=safe_json_dumps)
 
     @classmethod
     @EnodoAuth.auth.required
@@ -248,63 +196,8 @@ class ApiHandlers:
 
     @classmethod
     @EnodoAuth.auth.required
-    async def get_job_config_templates(cls, request):
-        """Get all job config templates from a specific job_type
-
-        Args:
-            request (Request): aiohttp request
-
-        Returns:
-            _type_: _description_
-        """
-        job_type = urllib.parse.unquote(
-            request.match_info['job_type'])
-        data, status = await BaseHandler.resp_get_job_config_templates(
-            job_type)
-        return web.json_response(
-            data=data, status=status)
-
-    @classmethod
-    @EnodoAuth.auth.required
-    async def add_job_config_templates(cls, request):
-        """Add a job config template
-
-        Args:
-            request (Request): aiohttp request
-
-        Returns:
-            _type_: _description_
-        """
-        try:
-            data = await request.json()
-        except JSONDecodeError as e:
-            resp, status = {'error': 'Invalid JSON'}, 400
-        else:
-            resp, status = await BaseHandler.resp_add_job_config_templates(
-                data)
-        return web.json_response(data=resp, status=status)
-
-    @classmethod
-    @EnodoAuth.auth.required
-    async def remove_job_config_templates(cls, request):
-        """Remove a job config template
-
-        Args:
-            request (Request): aiohttp request
-
-        Returns:
-            _type_: _description_
-        """
-        rid = urllib.parse.unquote(
-            request.match_info['rid'])
-        data, status = await BaseHandler.resp_remove_job_config_templates(
-            rid)
-        return web.json_response(
-            data=data, status=status)
-
-    @classmethod
-    @EnodoAuth.auth.required
-    async def get_series_config_templates(cls, request):
+    @implement_fields_query
+    async def get_series_config_templates(cls, request, fields=None):
         """Get all series config templates
 
         Args:
@@ -313,14 +206,6 @@ class ApiHandlers:
         Returns:
             _type_: _description_
         """
-        fields = None
-        if "fields" in request.rel_url.query:
-            fields = urllib.parse.unquote(
-                request.rel_url.query['fields'])
-            if fields == "":
-                fields = None
-            if fields is not None:
-                fields = fields.split(",")
         data, status = BaseHandler.resp_get_series_config_templates(
             fields=fields)
         return web.json_response(
@@ -366,7 +251,54 @@ class ApiHandlers:
 
     @classmethod
     @EnodoAuth.auth.required
-    async def get_enodo_event_outputs(cls, request):
+    async def update_series_config_templates_static(cls, request):
+        """Update a series config template
+
+        Args:
+            request (Request): aiohttp request
+
+        Returns:
+            _type_: _description_
+        """
+        try:
+            data = await request.json()
+        except JSONDecodeError as e:
+            resp, status = {'error': 'Invalid JSON'}, 400
+        rid = urllib.parse.unquote(
+            request.match_info['rid'])
+        data, status = \
+            await BaseHandler.resp_update_series_config_templates_static(
+                rid, data.get("name"), data.get("description"))
+        return web.json_response(
+            data=data, status=status)
+
+    @classmethod
+    @EnodoAuth.auth.required
+    async def update_series_config_templates(cls, request):
+        """Update a series config template
+
+        Args:
+            request (Request): aiohttp request
+
+        Returns:
+            _type_: _description_
+        """
+        try:
+            data = await request.json()
+        except JSONDecodeError as e:
+            resp, status = {'error': 'Invalid JSON'}, 400
+        rid = urllib.parse.unquote(
+            request.match_info['rid'])
+        data, status = \
+            await BaseHandler.resp_update_series_config_templates(
+                rid, data)
+        return web.json_response(
+            data=data, status=status)
+
+    @classmethod
+    @EnodoAuth.auth.required
+    @implement_fields_query
+    async def get_enodo_event_outputs(cls, request, fields=None):
         """Get all event outputs
 
         Args:
@@ -375,7 +307,7 @@ class ApiHandlers:
         Returns:
             _type_: _description_
         """
-        resp, status = await BaseHandler.resp_get_event_outputs()
+        resp, status = await BaseHandler.resp_get_event_outputs(fields=fields)
         return web.json_response(data=resp, status=status)
 
     @classmethod
@@ -423,7 +355,8 @@ class ApiHandlers:
 
     @classmethod
     @EnodoAuth.auth.required
-    async def get_possible_analyser_modules(cls, request):
+    @implement_fields_query
+    async def get_possible_analyser_modules(cls, request, fields=None):
         """Returns list of possible modules with corresponding parameters
 
         Args:
@@ -434,12 +367,14 @@ class ApiHandlers:
         """
 
         return web.json_response(
-            data=BaseHandler.resp_get_possible_analyser_modules(),
+            data=BaseHandler.resp_get_possible_analyser_modules(
+                fields=fields),
             status=200)
 
     @classmethod
     @EnodoAuth.auth.required
-    async def get_siridb_enodo_status(cls, request):
+    @implement_fields_query
+    async def get_siridb_enodo_status(cls, request, fields=None):
         """Get status of enodo hub
 
         Args:
@@ -448,7 +383,7 @@ class ApiHandlers:
         Returns:
             _type_: _description_
         """
-        resp = BaseHandler.resp_get_enodo_hub_status()
+        resp = BaseHandler.resp_get_enodo_hub_status(fields=fields)
         return web.json_response(data=resp, status=200)
 
     @classmethod
@@ -495,7 +430,8 @@ class ApiHandlers:
 
     @classmethod
     @EnodoAuth.auth.required
-    async def get_settings(cls, request):
+    @implement_fields_query
+    async def get_settings(cls, request, fields=None):
         """Returns current settings dict
 
         Args:
@@ -504,7 +440,7 @@ class ApiHandlers:
         Returns:
             _type_: _description_
         """
-        resp = BaseHandler.resp_get_enodo_config()
+        resp = BaseHandler.resp_get_enodo_config(fields=fields)
         return web.json_response(data=resp, status=200)
 
     @classmethod
@@ -549,7 +485,8 @@ class ApiHandlers:
 
     @classmethod
     @EnodoAuth.auth.required
-    async def get_enodo_stats(cls, request):
+    @implement_fields_query
+    async def get_enodo_stats(cls, request, fields=None):
         """Return stats and numbers from enodo domain
 
         Args:
@@ -558,9 +495,8 @@ class ApiHandlers:
         Returns:
             _type_: _description_
         """
-        resp = await BaseHandler.resp_get_enodo_stats()
-        return web.json_response(data={
-            'data': resp}, status=200)
+        resp = await BaseHandler.resp_get_enodo_stats(fields=fields)
+        return web.json_response(data=resp, status=200)
 
     @classmethod
     @EnodoAuth.auth.required