+++ lib/webserver/basehandler.py
@@ -1,4 +1,5 @@
 import logging
+from typing import Any
 from uuid import uuid4
 from aiohttp import web
 
@@ -27,17 +28,10 @@ from siridb.connector.lib.exceptions import (
     ServerError, UserAuthError)
 from lib.util.util import (
     get_job_config_output_series_name,
-    parse_output_series_name)
+    apply_fields_filter,
+    async_simple_fields_filter,
+    sync_simple_fields_filter)
 from version import VERSION
-
-
-def apply_fields_filter(resources: list, fields: list) -> list:
-    resources = [dict(t) for t in resources]
-    if fields is not None:
-        resources = [
-            {k: v for k, v in resource.items() if k in fields}
-            for resource in resources]
-    return resources
 
 
 class BaseHandler:
@@ -58,6 +52,7 @@ class BaseHandler:
             SeriesManager.get_all_series_names(regex_filter))}
 
     @classmethod
+    @async_simple_fields_filter(is_list=False, return_index=0)
     async def resp_get_single_monitored_series(cls, series_name):
         """Get monitored series details
 
@@ -69,13 +64,15 @@ class BaseHandler:
         """
         series = await SeriesManager.get_series_read_only(series_name)
         if series is None:
-            return {'data': ''}, 404
+            return {'data': {}}, 404
         series_data = series.to_dict()
         return {'data': series_data}, 200
 
     @classmethod
-    async def resp_get_all_series_output(cls, series_name, fields=None,
-                                         forecast_future_only=False):
+    async def resp_get_all_series_output(cls, series_name,
+                                         fields=None,
+                                         forecast_future_only=False,
+                                         types=None):
         """Get all series results
 
         Args:
@@ -88,22 +85,24 @@ class BaseHandler:
         if series is None:
             return web.json_response(data={'data': ''}, status=404)
 
-        should_fetch_data = True if fields is not None and \
-            "data" in fields else False
+        should_fetch_data = True if fields is None or \
+            (fields is not None and "data" in fields) else False
 
-        if should_fetch_data and not forecast_future_only:
+        data = None
+        if should_fetch_data and not forecast_future_only and types is None:
             data = await query_all_series_results(
                 ServerState.get_siridb_output_conn(), series_name)
 
         resp = []
-
         for job_config in series.config.job_config.values():
             if job_config.job_type == JOB_TYPE_BASE_SERIES_ANALYSIS:
                 continue  # Ignore since no output
+            if types is not None and job_config.job_type not in types:
+                continue  # Ignore due to filter
             points = None
             output_series_name = get_job_config_output_series_name(
                 series_name, job_config.job_type, job_config.config_name)
-            if should_fetch_data and not forecast_future_only:
+            if data is not None:
                 points = data.get(output_series_name)
             elif should_fetch_data:
                 points = (await SeriesManager.get_series_output_by_job_type(
@@ -111,62 +110,16 @@ class BaseHandler:
                     forecast_future_only=forecast_future_only)).get(
                         output_series_name)
             resp.append({
+                "output_series_name": output_series_name,
                 "config_name": job_config.config_name,
                 "job_type": job_config.job_type,
+                "job_meta": series.state.job_analysis_meta.get_job_meta(
+                    job_config.config_name),
                 "data": points
             })
         if fields is not None:
             resp = apply_fields_filter(resp, fields)
         return {'data': resp}
-
-    @classmethod
-    async def resp_get_series_forecasts(cls, series_name, only_future):
-        """Get series forecast results
-
-        Args:
-            series_name (string): name of series
-
-        Returns:
-            dict: dict with data
-        """
-        series = await SeriesManager.get_series_read_only(series_name)
-        if series is None:
-            return web.json_response(data={'data': ''}, status=404)
-        return {'data': await query_series_forecasts(
-            ServerState.get_siridb_output_conn(), series_name,
-            only_future=only_future)}
-
-    @classmethod
-    async def resp_get_series_anomalies(cls, series_name):
-        """Get series anomalies results
-
-        Args:
-            series_name (string): name of series
-
-        Returns:
-            dict: dict with data
-        """
-        series = await SeriesManager.get_series_read_only(series_name)
-        if series is None:
-            return web.json_response(data={'data': ''}, status=404)
-        return {'data': await query_series_anomalies(
-            ServerState.get_siridb_output_conn(), series_name)}
-
-    @classmethod
-    async def resp_get_series_static_rules_hits(cls, series_name):
-        """Get series static rules hits
-
-        Args:
-            series_name (string): name of series
-
-        Returns:
-            dict: dict with data
-        """
-        series = await SeriesManager.get_series_read_only(series_name)
-        if series is None:
-            return web.json_response(data={'data': ''}, status=404)
-        return {'data': await query_series_static_rules_hits(
-            ServerState.get_siridb_output_conn(), series_name)}
 
     @classmethod
     async def resp_run_siridb_query(cls, query):
@@ -190,13 +143,15 @@ class BaseHandler:
         return {'data': output}, 200
 
     @classmethod
+    @async_simple_fields_filter(return_index=0)
     async def resp_get_event_outputs(cls):
         """get all event output steams
 
         Returns:
             dict: dict with data
         """
-        return {'data': await EnodoEventManager.get_outputs()}, 200
+        outputs = await EnodoEventManager.get_outputs()
+        return {'data': outputs}, 200
 
     @classmethod
     async def resp_add_event_output(cls, output_type, data):
@@ -348,53 +303,10 @@ class BaseHandler:
                     if removed else 404
 
     @classmethod
-    async def resp_get_job_config_templates(cls, job_type):
-        trm = ResourceManager(
-            "job_config_templates", SeriesJobConfigTemplate, 0)
-        await trm.load()
-        templates = []
-        async for template in trm.itter():
-            if template.job_config_template.job_type == job_type:
-                templates.append(template)
-
-        return {'data': templates}, 200
-
-    @classmethod
-    async def resp_add_job_config_templates(cls, job_template: dict):
-        trm = ResourceManager(
-            "job_config_templates", SeriesJobConfigTemplate, 0)
-        await trm.load()
-        if trm.rid_exists(job_template.get("rid")):
-            return {'error': "template already exists"}, 400
-
-        if job_template.get("rid") is None:
-            job_template["rid"] = str(uuid4()).replace("-", "")
-
-        async with trm.create_resource(job_template) as template:
-            return {'data': template}, 201
-
-    @classmethod
-    async def resp_remove_job_config_templates(cls, rid: str):
-        trm = ResourceManager(
-            "job_config_templates", SeriesJobConfigTemplate, 0)
-        await trm.load()
-
-        if not trm.rid_exists(rid):
-            return {'error': "template does not exists"}, 404
-
-        template = await trm.get_resource(rid)
-        await trm.delete_resource(template)
-
-        return {'data': None}, 200
-
-    @classmethod
-    def resp_get_series_config_templates(cls, fields=None):
+    @sync_simple_fields_filter(return_index=0)
+    def resp_get_series_config_templates(cls):
         scrm = ServerState.series_config_template_rm
         templates = scrm.get_cached_resources()
-
-        # TODO: make this field filter general for all get endpoints
-        templates = apply_fields_filter(templates, fields)
-
         return {'data': templates}, 200
 
     @classmethod
@@ -412,15 +324,57 @@ class BaseHandler:
 
     @classmethod
     async def resp_remove_series_config_templates(cls, rid: str):
+        # TODO: check if series are using this template
+        scrm = ServerState.series_config_template_rm
+
+        if not scrm.rid_exists(rid):
+            return {'error': "template does not exists"}, 404
+
+        async for series in ServerState.series_rm.itter():
+            if series._config_from_template:
+                if series.config.rid == rid:
+                    return {'error': "template is used by series"}, 400
+
+        template = await scrm.get_resource(rid)
+        await scrm.delete_resource(template)
+
+        return {'data': None}, 200
+
+    @classmethod
+    async def resp_update_series_config_templates_static(cls, rid: str,
+                                                         name: str, desc: str):
         scrm = ServerState.series_config_template_rm
 
         if not scrm.rid_exists(rid):
             return {'error': "template does not exists"}, 404
 
         template = await scrm.get_resource(rid)
-        await scrm.delete_resource(template)
+        if name is not None:
+            template['name'] = name
+        if desc is not None:
+            template['description'] = desc
+        await template.store()
 
-        return {'data': None}, 200
+        return {'data': template}, 201
+
+    @classmethod
+    async def resp_update_series_config_templates(cls, rid: str,
+                                                  config: dict):
+        scrm = ServerState.series_config_template_rm
+
+        if not scrm.rid_exists(rid):
+            return {'error': "template does not exists"}, 404
+
+        template = await scrm.get_resource(rid)
+        if config is None:
+            return {'error', 'no config given', 400}
+        template['config'] = config
+        await template.store()
+        async for series in ServerState.series_rm.itter(update=True):
+            if series._config_from_template:
+                ServerState.index_series_schedules(series)
+
+        return {'data': template}, 201
 
     @classmethod
     def resp_get_jobs_queue(cls):
@@ -446,20 +400,22 @@ class BaseHandler:
                 series_name)}
 
     @classmethod
+    @sync_simple_fields_filter()
     def resp_get_possible_analyser_modules(cls):
         """Get all modules that are available
 
         Returns:
             dict: dict with data
         """
-        data = {'modules': list(ClientManager.modules.values())}
-        return {'data': data}
+        return {'data': list(ClientManager.modules.values())}
 
     @classmethod
+    @sync_simple_fields_filter(is_list=False)
     def resp_get_enodo_hub_status(cls):
         return {'data': {'version': VERSION}}
 
     @classmethod
+    @sync_simple_fields_filter(is_list=False)
     def resp_get_enodo_config(cls):
         return {'data': Config.get_settings(include_secrets=False)}
 
@@ -485,6 +441,7 @@ class BaseHandler:
         return {'data': True}
 
     @classmethod
+    @async_simple_fields_filter(is_list=False)
     async def resp_get_enodo_stats(cls):
         return {'data': {
             "no_series": SeriesManager.get_series_count(),
@@ -496,7 +453,7 @@ class BaseHandler:
                 "no_listeners": ClientManager.get_listener_count(),
                 "no_workers": ClientManager.get_worker_count(),
                 "no_busy_workers": ClientManager.get_busy_worker_count(),
-                "no_output_streams": len(EnodoEventManager.outputs)
+                "no_output_streams": len(await EnodoEventManager.get_outputs())
         }}
 
     @classmethod